Section 9 is done

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
     console.log(createdPost);
@@ -86,11 +87,19 @@ router.delete(
   checkAuth,
   (req, res, next) => {
   console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({
-      message: "Post deleted"
-  });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+
+    if (result.n > 0) {
+      res.status(200).json({
+        message: "Post deleted"
+       });
+    } else {
+      res.status(401).json({
+        message: "Not Authorized!"
+       });
+    }
+    // console.log(result);
+
   })
   .catch(() => {
 
@@ -111,14 +120,21 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
-    // console.log(result);
-    res.status(200).json({
-      message: "Successfully updated!"
-    })
+
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: "Successfully updated!"
+      });
+    } else {
+      res.status(401).json({
+        message: "Not Authorized"
+      });
+    }
+
   })
   .catch(() => {
